requirements.txt
pydantic>=2.0
python-dateutil==2.9.0.post0
black==25.1.0
ruff==0.12.8
Flask==3.1.1
Werkzeug==3.1.3
SQLAlchemy>=2.0
Flask-Migrate>=4.0
python-dotenv==1.1.1
psycopg2-binary==2.9.10
PyJWT>=2.7.0
openai==1.102.0
pgvector==0.2.5
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
sentence-transformers>=3.0.0
ollama>=0.1.8